Add and evaluate secondary config with increased noise.
Added another omnetpp config in which the noise is 3 dBm higher.
This simulates interference of the same level as the noise.
So an interfering message that is just weak enough to not be trained on.

Extended the Playground evaluation notebook to handle multiple runs.
Now we can read all results and show differences in the same plots.

Some texts in the notebook are outdated, though.